+++ bika/lims/browser/client/__init__.py
@@ -22,11 +22,9 @@ from views.batches import ClientBatchesView
 from views.analysisrequests import ClientAnalysisRequestsView
 from views.analysisprofiles import ClientAnalysisProfilesView
 from views.artemplates import ClientARTemplatesView
-from views.srtemplates import ClientSamplingRoundTemplatesView
 from views.samplepoints import ClientSamplePointsView
 from views.analysisspecs import ClientAnalysisSpecsView
 from views.attachments import ClientAttachmentsView
 from views.orders import ClientOrdersView
 from views.contacts import ClientContactsView
 from views.contacts import ClientContactVocabularyFactory
-from views.samplingrounds import ClientSamplingRoundsView